pyproject.toml
[project]
name = "llm-sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "ipykernel>=6.30.0",
    "jupyter>=1.1.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.97.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "sendgrid>=6.12.4",
]
